+++ CHANGELOG.md
@@ -9,6 +9,7 @@ License MIT https://raw.github.com/pontikis/dacapo/master/MIT_LICENSE
 
 Release 1.0.1 (12 Jun 2018)
 ---------------------------
+* Allow multiple connections #13
 * BUG FIX: setPgInsertSequence() must concern the current query - also renamed to setQueryInsertPgSequence() #12
 
 Release 1.0.0 (11 Jun 2018)